Refactor: js->ts
